DB: added account id parameter to getMapName, getMapResource, deleteMap, deleteMarker, deleteAnchor
git-svn-id: https://itsd.ict.sait.ca:9880/capstone07/trunk@266 0c6396e7-dfa7-ad41-aac3-ea9b0ac9ed18

+++ DB/Procedures/deleteMap.sql
@@ -3,18 +3,34 @@
 ********************************************************************* */
 
 
-DROP PROCEDURE IF EXISTS deleteMap;
+DROP FUNCTION IF EXISTS deleteMap;
 
 DELIMITER //
 
-CREATE PROCEDURE deleteMap 
-(p_map_id INT(4))
+CREATE FUNCTION deleteMap 
+(p_map_id INT(4), p_acc_id INT(4))
+RETURNS BOOLEAN
+READS SQL DATA
 BEGIN
-    CALL deleteAnchor(p_map_id);
-    CALL deleteMarker(p_map_id);
+    DECLARE flag BOOLEAN;
+    SET flag = false;
+    IF (SELECT COUNT(web_account_id)
+        FROM MAP
+        WHERE p_map_id = map_id
+        AND p_acc_id = web_account_id) != 0 THEN
+            SELECT deleteAnchor(p_map_id, p_acc_id) INTO flag;
+            SELECT deleteMarker(p_map_id, p_acc_id) INTO flag;
 
-    DELETE FROM MAP
-    WHERE map_id = p_map_id;
+            DELETE FROM MAP
+            WHERE map_id = p_map_id
+            AND p_acc_id = web_account_id;
+            
+            SET flag = true;
+    END IF;
+    RETURN flag;
 END//
 
 DELIMITER ;
+
+
+
